refactor: moved PrivateKey to circuits.js
