+++ yarn-project/circuits.js/src/types/partial_contract_address.ts
@@ -1,0 +1,6 @@
+import { Fr } from '../index.js';
+
+/**
+ * A type to represent a partially constructed contract address.
+ */
+export type PartialContractAddress = Fr;
