+++ yarn-project/circuits.js/src/types/public_key.ts
@@ -1,0 +1,4 @@
+import { Point } from '../index.js';
+
+/** Represents a user public key. */
+export type PublicKey = Point;
